+++ packages/terra-time-input/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Updated Jest Snapshots
+
 ## 4.41.0 - (July 20, 2021)
 
 * Changed